+++ server/src-lib/Hasura/GraphQL/Schema/BoolExp.hs
@@ -34,7 +34,8 @@ boolExp
   -> Maybe SelPermInfo
   -> m (Parser 'Input n (AnnBoolExp UnpreparedValue))
 boolExp table selectPermissions = memoizeOn 'boolExp table $ do
-  name <- qualifiedObjectToName table <&> (<> $$(G.litName "_bool_exp"))
+  tableGQLName <- getTableGQLName table
+  let name = tableGQLName <> $$(G.litName "_bool_exp")
   let description = G.Description $
         "Boolean expression to filter rows from the table " <> table <<>
         ". All fields are combined with a logical 'AND'."