+++ server/src-lib/Hasura/GraphQL/Schema/Select.hs
@@ -206,11 +206,11 @@ selectTableAggregate
 selectTableAggregate table fieldName description selectPermissions = runMaybeT do
   guard $ spiAllowAgg selectPermissions
   stringifyNum    <- asks $ qcStringifyNum . getter
-  tableName       <- lift $ qualifiedObjectToName table
+  tableGQLName    <- lift $ getTableGQLName table
   tableArgsParser <- lift $ tableArgs table selectPermissions
   aggregateParser <- lift $ tableAggregationFields table selectPermissions
   nodesParser     <- lift $ tableSelectionList table selectPermissions
-  let selectionName = tableName <> $$(G.litName "_aggregate")
+  let selectionName = tableGQLName <> $$(G.litName "_aggregate")
       aggregationParser = P.nonNullableParser $
         parsedSelectionsToFields RQL.TAFExp <$>
         P.selectionSet selectionName (Just $ G.Description $ "aggregated selection of " <>> table)
@@ -299,7 +299,7 @@ tableSelectionSet
   -> m (Parser 'Output n AnnotatedFields)
 tableSelectionSet table selectPermissions = memoizeOn 'tableSelectionSet table do
   tableInfo <- _tiCoreInfo <$> askTableInfo table
-  tableName <- qualifiedObjectToName table
+  tableGQLName <- getTableGQLName table
   let tableFields = Map.elems  $ _tciFieldInfoMap tableInfo
       tablePkeyColumns = _pkColumns <$> _tciPrimaryKey tableInfo
       description  = Just $ mkDescriptionWith (_tciDescription tableInfo) $
@@ -323,10 +323,10 @@ tableSelectionSet table selectPermissions = memoizeOn 'tableSelectionSet table d
             P.selection_ $$(G.litName "id") Nothing P.identifier $> RQL.AFNodeId table pkeyColumns
           allFieldParsers = fieldParsers <> [nodeIdFieldParser]
       nodeInterface <- node
-      pure $ P.selectionSetObject tableName description allFieldParsers [nodeInterface]
+      pure $ P.selectionSetObject tableGQLName description allFieldParsers [nodeInterface]
             <&> parsedSelectionsToFields RQL.AFExpression
     _                                 ->
-      pure $ P.selectionSetObject tableName description fieldParsers []
+      pure $ P.selectionSetObject tableGQLName description fieldParsers []
             <&> parsedSelectionsToFields RQL.AFExpression
 
 -- | List of table fields object.
@@ -377,9 +377,9 @@ tableConnectionSelectionSet
   -> SelPermInfo
   -> m (Parser 'Output n (RQL.ConnectionFields UnpreparedValue))
 tableConnectionSelectionSet table selectPermissions = do
-  tableName   <- qualifiedObjectToName table
-  edgesParser <- tableEdgesSelectionSet
-  let connectionTypeName = tableName <> $$(G.litName "Connection")
+  edgesParser  <- tableEdgesSelectionSet
+  tableGQLName <- getTableGQLName table
+  let connectionTypeName = tableGQLName <> $$(G.litName "Connection")
       pageInfo = P.subselection_ $$(G.litName "pageInfo") Nothing
                  pageInfoSelectionSet <&> RQL.ConnectionPageInfo
       edges = P.subselection_ $$(G.litName "edges") Nothing
@@ -406,12 +406,13 @@ tableConnectionSelectionSet table selectPermissions = do
       in P.nonNullableParser $ P.selectionSet $$(G.litName "PageInfo") Nothing allFields
          <&> parsedSelectionsToFields RQL.PageInfoTypename
 
+
     tableEdgesSelectionSet
       :: m (Parser 'Output n (RQL.EdgeFields UnpreparedValue))
     tableEdgesSelectionSet = do
-      tableName           <- qualifiedObjectToName table
+      tableGQLName        <- getTableGQLName table
       edgeNodeParser      <- P.nonNullableParser <$> tableSelectionSet table selectPermissions
-      let edgesType = tableName <> $$(G.litName "Edge")
+      let edgesType = tableGQLName <> $$(G.litName "Edge")
           cursor    = P.selection_ $$(G.litName "cursor") Nothing
                       P.string $> RQL.EdgeCursor
           edgeNode  = P.subselection_ $$(G.litName "node") Nothing
@@ -454,10 +455,11 @@ selectFunctionAggregate function fieldName description selectPermissions = runMa
   let table = fiReturnType function
   stringifyNum <- asks $ qcStringifyNum . getter
   guard $ spiAllowAgg selectPermissions
+  tableGQLName <- getTableGQLName table
   tableArgsParser    <- lift $ tableArgs table selectPermissions
   functionArgsParser <- lift $ customSQLFunctionArgs function
   aggregateParser    <- lift $ tableAggregationFields table selectPermissions
-  selectionName      <- lift $ qualifiedObjectToName table <&> (<> $$(G.litName "_aggregate"))
+  selectionName      <- lift $ pure tableGQLName <&> (<> $$(G.litName "_aggregate"))
   nodesParser        <- lift $ tableSelectionList table selectPermissions
   let argsParser = liftA2 (,) functionArgsParser tableArgsParser
       aggregationParser = fmap (parsedSelectionsToFields RQL.TAFExp) $
@@ -784,11 +786,11 @@ tableAggregationFields
   -> SelPermInfo
   -> m (Parser 'Output n RQL.AggregateFields)
 tableAggregationFields table selectPermissions = do
-  tableName  <- qualifiedObjectToName table
+  tableGQLName  <- getTableGQLName table
   allColumns <- tableSelectColumns table selectPermissions
   let numericColumns   = onlyNumCols allColumns
       comparableColumns  = onlyComparableCols allColumns
-      selectName   = tableName <> $$(G.litName "_aggregate_fields")
+      selectName   = tableGQLName <> $$(G.litName "_aggregate_fields")
       description  = G.Description $ "aggregate fields of " <>> table
   count     <- countField
   numericAndComparable <- fmap concat $ sequenceA $ catMaybes
@@ -796,12 +798,12 @@ tableAggregationFields table selectPermissions = do
       if null numericColumns then Nothing else Just $
         for numericAggOperators $ \operator -> do
           numFields <- mkNumericAggFields operator numericColumns
-          pure $ parseAggOperator operator tableName numFields
+          pure $ parseAggOperator operator tableGQLName numFields
     , -- operators on comparable columns
       if null comparableColumns then Nothing else Just $ do
         comparableFields <- traverse mkColumnAggField comparableColumns
         pure $ comparisonAggOperators & map \operator ->
-               parseAggOperator operator tableName comparableFields
+               parseAggOperator operator tableGQLName comparableFields
     ]
   let aggregateFields = count : numericAndComparable
   pure $ P.selectionSet selectName (Just description) aggregateFields
@@ -840,9 +842,9 @@ tableAggregationFields table selectPermissions = do
       -> G.Name
       -> [FieldParser n RQL.PGColFld]
       -> FieldParser n RQL.AggregateField
-    parseAggOperator operator tableName columns =
+    parseAggOperator operator tableGQLName columns =
       let opText  = G.unName operator
-          setName = tableName <> $$(G.litName "_") <> operator <> $$(G.litName "_fields")
+          setName = tableGQLName <> $$(G.litName "_") <> operator <> $$(G.litName "_fields")
           setDesc = Just $ G.Description $ "aggregate " <> opText <> " on columns"
           subselectionParser = P.selectionSet setName setDesc columns
             <&> parsedSelectionsToFields RQL.PCFExp