+++ server/src-lib/Hasura/RQL/DDL/Schema/Table.hs
@@ -17,6 +17,9 @@ module Hasura.RQL.DDL.Schema.Table
   , SetTableCustomFields(..)
   , runSetTableCustomFieldsQV2
 
+  , SetTableCustomization(..)
+  , runSetTableCustomization
+
   , buildTableCache
   , delTableAndDirectDeps
   , processTableChanges
@@ -196,6 +199,13 @@ runSetExistingTableIsEnumQ (SetTableIsEnum tableName isEnum) = do
   buildSchemaCacheFor (MOTable tableName)
   return successMsg
 
+data SetTableCustomization
+  = SetTableCustomization
+  { _stcTable         :: !QualifiedTable
+  , _stcConfiguration :: !TableConfig
+  } deriving (Show, Eq, Lift)
+$(deriveJSON (aesonDrop 4 snakeCase) ''SetTableCustomization)
+
 data SetTableCustomFields
   = SetTableCustomFields
   { _stcfTable             :: !QualifiedTable
@@ -215,8 +225,17 @@ runSetTableCustomFieldsQV2
   :: (MonadTx m, CacheRWM m) => SetTableCustomFields -> m EncJSON
 runSetTableCustomFieldsQV2 (SetTableCustomFields tableName rootFields columnNames) = do
   void $ askTabInfo tableName -- assert that table is tracked
-  updateTableConfig tableName (TableConfig rootFields columnNames)
+  -- `Identifier` is set to `Nothing` below because this API doesn't accept it
+  updateTableConfig tableName (TableConfig rootFields columnNames Nothing)
   buildSchemaCacheFor (MOTable tableName)
+  return successMsg
+
+runSetTableCustomization
+  :: (MonadTx m, CacheRWM m) => SetTableCustomization -> m EncJSON
+runSetTableCustomization (SetTableCustomization table config) = do
+  void $ askTabInfo table
+  updateTableConfig table config
+  buildSchemaCacheFor (MOTable table)
   return successMsg
 
 unTrackExistingTableOrViewP1
@@ -271,7 +290,8 @@ processTableChanges ti tableDiff = do
         procAlteredCols sc tn
 
       withNewTabName newTN = do
-        let tnGQL = snakeCaseQualObject newTN
+        let customTableNameText = G.unName <$> (_tcCustomName . _tciCustomConfig $ ti)
+            tnGQL = fromMaybe (snakeCaseQualObject newTN) customTableNameText
         -- check for GraphQL schema conflicts on new name
         checkConflictingNode sc tnGQL
         procAlteredCols sc tn
@@ -287,10 +307,10 @@ processTableChanges ti tableDiff = do
     TableDiff mNewName droppedCols _ alteredCols _ computedFieldDiff _ _ = tableDiff
 
     possiblyDropCustomColumnNames tn = do
-      let TableConfig customFields customColumnNames = _tciCustomConfig ti
+      let TableConfig customFields customColumnNames identifier = _tciCustomConfig ti
           modifiedCustomColumnNames = foldl' (flip Map.delete) customColumnNames droppedCols
       when (modifiedCustomColumnNames /= customColumnNames) $
-        liftTx $ updateTableConfig tn $ TableConfig customFields modifiedCustomColumnNames
+        liftTx $ updateTableConfig tn $ TableConfig customFields modifiedCustomColumnNames identifier
 
     procAlteredCols sc tn = for_ alteredCols $
       \( PGRawColumnInfo oldName _ oldType _ _