+++ server/src-lib/Hasura/RQL/Types/Table.hs
@@ -378,20 +378,22 @@ data TableConfig
   = TableConfig
   { _tcCustomRootFields  :: !TableCustomRootFields
   , _tcCustomColumnNames :: !CustomColumnNames
+  , _tcCustomName        :: !(Maybe G.Name)
   } deriving (Show, Eq, Lift, Generic)
 instance NFData TableConfig
 instance Cacheable TableConfig
-$(deriveToJSON (aesonDrop 3 snakeCase) ''TableConfig)
+$(deriveToJSON (aesonDrop 3 snakeCase){omitNothingFields=True} ''TableConfig)
 
 emptyTableConfig :: TableConfig
 emptyTableConfig =
-  TableConfig emptyCustomRootFields M.empty
+  TableConfig emptyCustomRootFields M.empty Nothing
 
 instance FromJSON TableConfig where
   parseJSON = withObject "TableConfig" $ \obj ->
     TableConfig
     <$> obj .:? "custom_root_fields" .!= emptyCustomRootFields
     <*> obj .:? "custom_column_names" .!= M.empty
+    <*> obj .:? "custom_name"
 
 -- | The @field@ and @primaryKeyColumn@ type parameters vary as the schema cache is built and more
 -- information is accumulated. See 'TableRawInfo' and 'TableCoreInfo'.
@@ -421,7 +423,8 @@ type TableCoreInfo = TableCoreInfoG FieldInfo PGColumnInfo
 
 tciUniqueOrPrimaryKeyConstraints :: TableCoreInfoG a b -> Maybe (NonEmpty Constraint)
 tciUniqueOrPrimaryKeyConstraints info = NE.nonEmpty $
-  maybeToList (_pkConstraint <$> _tciPrimaryKey info) <> toList (_tciUniqueConstraints info)
+  maybeToList (_pkConstraint <$> _tciPrimaryKey info)
+  <> toList (_tciUniqueConstraints info)
 
 data TableInfo
   = TableInfo