+++ server/src-lib/Hasura/SQL/Types.hs
@@ -52,6 +52,7 @@ module Hasura.SQL.Types
   , qualObjectToText
   , snakeCaseQualObject
   , qualifiedObjectToName
+  , isGraphQLCompliantTableName
 
   , PGScalarType(..)
   , WithScalarType(..)
@@ -319,6 +320,9 @@ qualifiedObjectToName objectName = do
   onNothing (G.mkName textName) $ throw400 ValidationFailed $
     "cannot include " <> objectName <<> " in the GraphQL schema because " <> textName
     <<> " is not a valid GraphQL identifier"
+
+isGraphQLCompliantTableName :: ToTxt a => QualifiedObject a -> Bool
+isGraphQLCompliantTableName = isJust . G.mkName . snakeCaseQualObject
 
 type QualifiedTable = QualifiedObject TableName