+++ server/src-lib/Hasura/Server/API/Query.hs
@@ -129,11 +129,12 @@ data RQLQueryV1
   | RQDumpInternalState !DumpInternalState
 
   | RQSetCustomTypes !CustomTypes
+  | RQSetTableCustomization !SetTableCustomization
   deriving (Show, Eq)
 
 data RQLQueryV2
   = RQV2TrackTable !TrackTableV2
-  | RQV2SetTableCustomFields !SetTableCustomFields
+  | RQV2SetTableCustomFields !SetTableCustomFields -- deprecated
   | RQV2TrackFunction !TrackFunctionV2
   deriving (Show, Eq)
 
@@ -304,6 +305,7 @@ queryModifiesSchemaCache (RQV1 qi) = case qi of
 
   RQDumpInternalState _           -> False
   RQSetCustomTypes _              -> True
+  RQSetTableCustomization _       -> True
 
   RQBulk qs                       -> any queryModifiesSchemaCache qs
 
@@ -440,6 +442,7 @@ runQueryM env rq = withPathK "args" $ case rq of
       RQRunSql q                   -> runRunSQL q
 
       RQSetCustomTypes q           -> runSetCustomTypes q
+      RQSetTableCustomization q    -> runSetTableCustomization q
 
       RQBulk qs                    -> encJFromList <$> indexedMapM (runQueryM env) qs
 
@@ -528,6 +531,7 @@ requiresAdmin = \case
 
     RQDumpInternalState _           -> True
     RQSetCustomTypes _              -> True
+    RQSetTableCustomization _       -> True
 
     RQRunSql _                      -> True