+++ server/src-lib/Hasura/Server/Migrate.hs
@@ -354,7 +354,7 @@ recreateSystemMetadata = do
     for_ tableRels \case
       Left relDef -> insertRelationshipToCatalog tableName ObjRel relDef
       Right relDef -> insertRelationshipToCatalog tableName ArrRel relDef
-  buildSchemaCacheStrict
+  buildSchemaCache
   where
     systemMetadata :: [(QualifiedTable, [Either ObjRelDef ArrRelDef])]
     systemMetadata =